Remove usages of Key one by one. Move KeyUntil into encrption api module
