+++ config/src/main/java/com/quorum/tessera/config/keys/KeyEncryptorImpl.java
@@ -4,9 +4,9 @@ import com.quorum.tessera.argon2.Argon2;
 import com.quorum.tessera.argon2.ArgonResult;
 import com.quorum.tessera.config.ArgonOptions;
 import com.quorum.tessera.config.PrivateKeyData;
+import com.quorum.tessera.encryption.KeyUtil;
 import com.quorum.tessera.encryption.PrivateKey;
 import com.quorum.tessera.encryption.SharedKey;
-import com.quorum.tessera.nacl.Key;
 import com.quorum.tessera.nacl.NaclFacade;
 import com.quorum.tessera.nacl.Nonce;
 import org.slf4j.Logger;
@@ -57,18 +57,18 @@ public class KeyEncryptorImpl implements KeyEncryptor {
 
         final ArgonResult argonResult;
 
-        if(argonOptions == null) {
+        if (argonOptions == null) {
             argonResult = this.argon2.hash(password, salt);
         } else {
             argonResult = this.argon2.hash(
-                new com.quorum.tessera.argon2.ArgonOptions(
-                    argonOptions.getAlgorithm(),
-                    argonOptions.getIterations(),
-                    argonOptions.getMemory(),
-                    argonOptions.getParallelism()
-                ),
-                password,
-                salt
+                    new com.quorum.tessera.argon2.ArgonOptions(
+                            argonOptions.getAlgorithm(),
+                            argonOptions.getIterations(),
+                            argonOptions.getMemory(),
+                            argonOptions.getParallelism()
+                    ),
+                    password,
+                    salt
             );
         }
 
@@ -76,7 +76,7 @@ public class KeyEncryptorImpl implements KeyEncryptor {
         LOGGER.debug("Generated the random nonce {}", nonce);
 
         final byte[] encryptedKey = this.nacl.sealAfterPrecomputation(
-            privateKey.getKeyBytes(), nonce, SharedKey.from(argonResult.getHash())
+                privateKey.getKeyBytes(), nonce, SharedKey.from(argonResult.getHash())
         );
 
         LOGGER.info("Private key encrypted");
@@ -86,22 +86,22 @@ public class KeyEncryptorImpl implements KeyEncryptor {
         final String sbox = this.encoder.encodeToString(encryptedKey);
 
         return new PrivateKeyData(
-            null,
-            snonce,
-            asalt,
-            sbox,
-            new ArgonOptions(
-                argonResult.getOptions().getAlgorithm(),
-                argonResult.getOptions().getIterations(),
-                argonResult.getOptions().getMemory(),
-                argonResult.getOptions().getParallelism()
-            ),
-            null
+                null,
+                snonce,
+                asalt,
+                sbox,
+                new ArgonOptions(
+                        argonResult.getOptions().getAlgorithm(),
+                        argonResult.getOptions().getIterations(),
+                        argonResult.getOptions().getMemory(),
+                        argonResult.getOptions().getParallelism()
+                ),
+                null
         );
     }
 
     @Override
-    public Key decryptPrivateKey(final PrivateKeyData privateKey, final String password) {
+    public PrivateKey decryptPrivateKey(final PrivateKeyData privateKey, final String password) {
 
         LOGGER.info("Decrypting private key");
         LOGGER.debug("Decrypting private key {} using password {}", privateKey.getValue(), password);
@@ -109,26 +109,28 @@ public class KeyEncryptorImpl implements KeyEncryptor {
         final byte[] salt = this.decoder.decode(privateKey.getAsalt());
 
         final ArgonResult argonResult = this.argon2.hash(
-            new com.quorum.tessera.argon2.ArgonOptions(
-                privateKey.getArgonOptions().getAlgorithm(),
-                privateKey.getArgonOptions().getIterations(),
-                privateKey.getArgonOptions().getMemory(),
-                privateKey.getArgonOptions().getParallelism()
-            ),
-            password,
-            salt
+                new com.quorum.tessera.argon2.ArgonOptions(
+                        privateKey.getArgonOptions().getAlgorithm(),
+                        privateKey.getArgonOptions().getIterations(),
+                        privateKey.getArgonOptions().getMemory(),
+                        privateKey.getArgonOptions().getParallelism()
+                ),
+                password,
+                salt
         );
 
         final byte[] originalKey = this.nacl.openAfterPrecomputation(
-            this.decoder.decode(privateKey.getSbox()),
-            new Nonce(this.decoder.decode(privateKey.getSnonce())),
-            SharedKey.from(argonResult.getHash())
+                this.decoder.decode(privateKey.getSbox()),
+                new Nonce(this.decoder.decode(privateKey.getSnonce())),
+                SharedKey.from(argonResult.getHash())
         );
 
-        LOGGER.info("Decrypted private key");
-        LOGGER.debug("Decrypted private key {}", new Key(originalKey));
+        PrivateKey outcome = PrivateKey.from(originalKey);
 
-        return new Key(originalKey);
+        LOGGER.info("Decrypted private key");
+        LOGGER.debug("Decrypted private key {}", KeyUtil.encodeToBase64(outcome));
+
+        return outcome;
     }
 
 }
