+++ tessera-core/src/main/java/com/quorum/tessera/key/KeyManagerImpl.java
@@ -1,5 +1,6 @@
 package com.quorum.tessera.key;
 
+import com.quorum.tessera.encryption.KeyUtil;
 import com.quorum.tessera.encryption.PrivateKey;
 import com.quorum.tessera.encryption.PublicKey;
 import com.quorum.tessera.config.keypairs.ConfigKeyPair;
